+++ src/index.js
@@ -5,13 +5,14 @@ import App from "./components/App";
 import * as serviceWorker from "./serviceWorker";
 import { Provider } from "react-redux";
 import store from './store';
-import { BrowserRouter as Router } from 'react-router-dom';
+import {
+  HashRouter as Router
+} from "react-router-dom";
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-
-      <Router>
+      <Router  basename={process.env.PUBLIC_URL}>
         <App />
       </Router>
     </Provider>